docs: udpated examples
